requirements.txt
llama-index==0.9.4
streamlit== 1.32.2
openai== 1.3.4
python-dotenv==1.0.1